requirements.txt
aioredis==1.3.1
alabaster==0.7.12
amqp==2.6.1
appdirs==1.4.4
argon2-cffi==19.1.0
asgiref==3.3.4
astroid==2.4.2
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
autobahn==21.2.1
Automat==20.2.0
Babel==2.9.0
backcall==0.2.0
billiard==3.6.3.0
black==19.3b0
celery==4.3.0
certifi==2020.12.5
cffi==1.14.4
channels==2.4.0
channels-redis==3.2.0
chardet==4.0.0
click==7.1.2
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
cryptography==3.4.6
daphne==2.5.0
decorator==4.4.2
defusedxml==0.6.0
Django==2.2.3
django-allauth==0.39.1
django-appconf==1.0.4
django-celery-beat==1.5.0
django-celery-email==3.0.0
django-compressor==2.3
django-contrib-comments==2.0.0
django-coverage-plugin==1.6.0
django-crispy-forms==1.7.2
django-debug-toolbar==2.0
django-environ==0.4.5
django-extensions==2.1.9
django-haystack==3.0
django-markdownx==3.0.1
django-model-utils==3.2.0
django-redis==4.10.0
django-simpleui==2021.4.1
django-taggit==1.3.0
django-test-plus==1.4.0
django-timezone-field==4.1.1
djangorestframework==3.9.4
docutils==0.16
elasticsearch==2.4.1
entrypoints==0.3
factory-boy==2.12.0
Faker==2.0.1
flake8==3.7.7
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
imagesize==1.2.0
importlib-metadata==3.3.0
incremental==21.3.0
ipdb==0.12
ipython==7.16.1
ipython-genutils==0.2.0
isort==5.6.4
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
kombu==4.6.11
lazy-object-proxy==1.4.3
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.6.0
msgpack==1.0.2
msgpack-python==0.5.6
mypy==0.711
mypy-extensions==0.4.3
mysqlclient==1.4.2.post1
oauthlib==3.1.0
packaging==20.8
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.0.0
pluggy==0.13.1
prompt-toolkit==3.0.8
psycopg2-binary==2.8.3
ptyprocess==0.6.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
Pygments==2.7.3
pylint==2.6.0
pylint-celery==0.3
pylint-django==2.0.9
pylint-plugin-utils==0.6
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==5.0.0
pytest-django==3.5.1
pytest-sugar==0.9.2
python-crontab==2.5.1
python-dateutil==2.8.1
python-slugify==3.0.2
python3-openid==3.2.0
pytz==2019.1
rcssmin==1.0.6
redis==3.2.1
requests==2.25.1
requests-oauthlib==1.3.0
rjsmin==1.1.0
service-identity==18.1.0
six==1.15.0
slugify==0.0.1
snowballstemmer==2.0.0
sorl-thumbnail==12.7.0
Sphinx==2.1.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
termcolor==1.1.0
text-unidecode==1.2
toml==0.10.2
traitlets==4.3.3
Twisted==21.2.0
txaio==21.2.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
vine==1.3.0
wcwidth==0.2.5
Werkzeug==0.14.1
wrapt==1.12.1
zipp==3.4.0
zope.interface==5.2.0
uwsgi
